Add loci query algorithm
This algorithm queries each k-mer in the graph to GCSA index. It counts
the number of loci where at least one k-mer are not present in the GCSA
index. However, it doesn't output the exact number of uncovered loci
because for each k-mer query which are present in the GCSA, the list of
occurrences should also be verified to make sure whether the current
locus is in the list or not.